Remove duplicates of your own public entries
OK. Moving on to the next problem: the wrong entry ID
seems to get passed when I click a thing. this results in it thinking
I'm updating pincho's entry when I'm updating Pixie.

These form controls should have IDs that are unique to the page.
That might be its problem.